+++ authlib/google.py
@@ -21,10 +21,7 @@ class GoogleOAuth2Client(OAuthClient):
         self._session = OAuth2Session(
             self.client_id,
             scope=self.scope,
-            redirect_uri=request.build_absolute_uri(reverse(
-                'accounts_oauth2',
-                kwargs={'provider': 'google'},
-            )),
+            redirect_uri=request.build_absolute_uri('.'),
         )
 
     def get_authentication_url(self):